+++ app/Http/Controllers/Admin/TicketController.php
@@ -11,8 +11,11 @@ use Illuminate\Support\Str;
 
 use App\Models\CustomUserTable;
 use App\Models\AdminProfile;
+use App\Models\UserProfile;
 use App\Models\Ticket;
+use App\Models\TicketComment;
 use App\Models\Office;
+use App\Models\AuditTrail;
 
 class TicketController extends Controller
 {
@@ -78,6 +81,141 @@ class TicketController extends Controller
             }
         }
         else {
+            // If the User is Anonymous
+            return response()->json([
+                'status' => 'error',
+                'message' => "Unauthorized Access.",
+            ], 409);
+        }
+    }
+
+    // View Ticket
+    // Get Ticket Info
+    public function backend_getTicketInfo(Request $request, $ticket_number)
+    {
+        if (Auth::check()) {
+            // Check if the user is not an admin
+            if (!Auth::user()->is_admin) {
+                return response()->json([
+                    'status' => 'error',
+                    'message' => "Unauthorized Access.",
+                ], 409);
+            } else {
+                $ticket = Ticket::where('ticket_number', $ticket_number)->first();
+
+                $ticketdata = [];
+
+                if ($ticket) {
+
+                    $office = Office::where('office_id', $ticket->office_id)->first();
+                    $user = UserProfile::where('user_id', $ticket->user_id)->first();
+
+                    $ticketData[] = [
+                        'ticket_id' => $ticket->ticket_id,
+                        'affiliation' => $ticket->affiliation,
+                        'created_at' => $ticket->created_at,
+                        'ticket_number' => $ticket->ticket_number,
+                        'priority' => $ticket->priority,
+                        'status' => $ticket->status,
+                        'type' => $ticket->type,
+                        'office' => $office->office_name,
+                        'service' => $ticket->service,
+                        'resolved_date' => $ticket->resolved_date,
+                        'user' => $user->last_name." ".$user->first_name,
+                    ];
+
+                    // admin Verification
+                    $admin = Auth::user();
+                    $adminProfile = AdminProfile::where('user_id', $admin->user_id)->first();
+
+                    if ($adminProfile->office_id == $ticket->office_id) {
+                        return response()->json([
+                            'status' => 'success',
+                            'message' => "Access Granted.",
+                            'data' => $ticketData,
+                            'verify' => 'matched',
+                        ], 200);
+                    }
+                    else {
+                        return response()->json([
+                            'status' => 'error',
+                            'verify' => 'unmatched',
+                            'message' => "Access Denied.",
+                        ], 200);
+                    }
+
+                }
+                else {
+                    return response()->json([
+                        'status' => 'error',
+                        'message' => "Data not found.",
+                    ], 404);
+                }
+            }
+        }
+        else 
+        {
+            // If the User is Anonymous
+            return response()->json([
+                'status' => 'error',
+                'message' => "Unauthorized Access.",
+            ], 409);
+        }
+    }
+
+    public function backend_getTicketComment(Request $request, $ticket_number)
+    {
+        if (Auth::check()) {
+            // Check if the user not an admin
+            if (!Auth::user()->is_admin) {
+                return response()->json([
+                    'status' => 'error',
+                    'message' => "Unauthorized Access.",
+                ], 409);
+            } else {
+                $user = Auth::user();
+                $ticket = Ticket::where('ticket_number', $ticket_number)->first();
+                $comment = TicketComment::where('ticket_id', $ticket->ticket_id)->orderBy('created_at', 'asc')->get();
+
+                return response()->json([
+                    'status' => 'success',
+                    'message' => "Access Granted.",
+                    'user_id' => $user->user_id,
+                    'data' => $comment,
+                ], 200);
+            }
+        }
+        else 
+        {
+            // If the User is Anonymous
+            return response()->json([
+                'status' => 'error',
+                'message' => "Unauthorized Access.",
+            ], 409);
+        }
+    }
+
+    public function backend_getAuditTrail(Request $request)
+    {
+        if (Auth::check()) {
+            // Check if the user is not an admin
+            if (!Auth::user()->is_admin) {
+                return response()->json([
+                    'status' => 'error',
+                    'message' => "Unauthorized Access.",
+                ], 409);
+            } else {
+                $trail = AuditTrail::where('reference', $request->ticket_number)->orderBy('created_at', 'asc')->get();
+
+                return response()->json([
+                    'status' => 'success',
+                    'message' => "Access Granted.",
+                    'data' => $trail,
+                ], 200);
+            }
+        }
+        else 
+        {
             // If the User is Anonymous
             return response()->json([
                 'status' => 'error',