+++ routes/userbackend.php
@@ -18,4 +18,5 @@ Route::get('/backend/user/getOffice', [TicketController::class, 'backend_getOffi
 // View Ticket
 Route::get('/backend/user/getTicketInfo/{ticket_number}', [TicketController::class, 'backend_getTicketInfo']);
 Route::get('/backend/user/getTicketComment/{ticket_number}', [TicketController::class, 'backend_getTicketComment']);
+Route::get('/backend/user/getAuditTrails/{ticket_number}', [TicketController::class, 'backend_getAuditTrail']);
 Route::post('/backend/user/addTicketComment', [TicketController::class, 'backend_addTicketComment']);